pdf-engine/requirements.txt
pypdf
pymupdf
pillow
numpy>=4.2.0
pdf2image>=1.17.0
Pillow
reportlab>=4.1.0
cryptography>=42.0.5
pdf2docx
